Share skill setup file exclusion rules
Centralize generated skill file exclusions so layout discovery, directory sync, and cleanup paths use the same .meta, .DS_Store, and .gitkeep rules.
