+++ Packages/src/Editor/Infrastructure/SkillSetup/SkillSetupFileExclusion.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bd80275d3f2e446a8d15695ff85aab7b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
